good joke before, about it working. This time tho
