+++ Dockerfile
@@ -20,6 +20,7 @@ ADD package.json package-lock.json ./
 RUN npm prune --production
 
 COPY --from=build /app/build ./build
+ADD index.prod.js ./
 
 COPY kysely.config.ts ./
 COPY migrations ./migrations
